Fix Ractor safety
Unshareable objects can't be sent to ractors as today these objects are
unsafe to concurrently modify. In the future (under Ractor-local GC)
this will be even more broken and reading from the objects may also
break (as they could be GC'd).

This enforces shareability in the C extension, and adjusts tests to work
using some Ruby 3.5 specific features. Some of the patterns here
(sharing a proc) I don't know of a way to make work on Ruby 3.4 and I
don't think we should bother.